Move cute-specific logic from runtime/__init__.py into runtime/cute
The cutedsl (cutlass) launch path -- ~3.6k lines of wrapper codegen, tcgen05
grouped-scheduler metadata, launch-arg caching, on-disk compile cache, and the
CUDA-graph integration -- lived inline in `helion/runtime/__init__.py`. Move it
into a dedicated backend subpackage so the package root stays backend-neutral,
matching `runtime/pallas/` and `runtime/triton/`:

    runtime/__init__.py (cute block) -> runtime/cute/launcher.py

`runtime/__init__.py` keeps the backend-neutral helpers (`default_launcher`,
`get_num_sm`, `default_metal_launcher`) and re-exports the cute public API
(`default_cute_launcher`, `cute_cuda_graph`) plus every private helper the tests
and generated code read via `helion.runtime.<name>`, so `from helion.runtime
import ...` and attribute access keep working.

The moved code is byte-identical apart from the relative-import depth shifting by
one (e.g. `.._compiler.cute` -> `..._compiler.cute`); `get_num_sm` now comes
straight from `..triton.launcher` to avoid a circular import (identical behaviour
for the CUDA-only cute path). `cutlass` stays a lazy in-function import, so
importing the module never requires a CuTe install.

`test_cute_backend.py` patches the launch internals; those `patch(...)` targets
move to `helion.runtime.cute.launcher.<fn>` -- a mock must patch where the name
is looked up, and the call sites now live in that module (a re-export in
`helion.runtime` would not be intercepted). Direct reads of the private helpers
(e.g. `helion_runtime._CuteLaunchArgCacheEntry`) instead rely on the re-exports.

Verified on a Blackwell (B200) machine with cutlass installed: the previously
failing 23 `test_cute_backend.py` cases (all `AttributeError` on moved symbols)
now pass, and the full-suite failure set is byte-identical to the pre-refactor
parent commit -- the refactor changes zero test outcomes.

stack-info: PR: https://github.com/pytorch/helion/pull/3228, branch: AmesingFlank/stack/107
